Remove old contour API
